Make GH Action use mermaid-cli as base for Docker image

+++ render-md-mermaid.sh
@@ -68,10 +68,15 @@ for mermaid_img in $(find . -name "*.mermaid" | sed -E 's/((.*).mermaid)/\2|\1/'
 do
     image_file=${mermaid_img%|*}
     mermaid_file=${mermaid_img#*|}
-    docker run --rm -t -v "$PWD:/data" minlag/mermaid-cli:latest -o "/data/$image_file" -i "/data/$mermaid_file" -t neutral -C "/data/.render-md-mermaid.css" -c "/data/.render-md-mermaid-config.json" -s 4
+    if [[ "$2" == "in-container" ]]; then
+        /usr/local/bin/mmdc -p /puppeteer-config.json -o "$image_file" -i "$mermaid_file" -t neutral -C ".render-md-mermaid.css" -c ".render-md-mermaid-config.json" -s 4
+    else
+        docker run --rm -t -v "$PWD:/data" minlag/mermaid-cli:latest -o "/data/$image_file" -i "/data/$mermaid_file" -t neutral -C "/data/.render-md-mermaid.css" -c "/data/.render-md-mermaid-config.json" -s 4
+    fi
     if [[ "$image_file" =~ ^.*\.svg$ ]]; then
         sed -i.bak -e 's/<br>/<br\/>/g' $image_file
     fi
+    echo "Generated: $image_file"
     rm -f "$mermaid_file" "$image_file.bak"
 done
 rm -f .render-md-mermaid-config.json .render-md-mermaid.css
